Rename FilterQuery to StringQuery
FilterQuery still carries the old name used when this library was still a Bundle,
but the name tells nothing. This library is about searching, not filtering.

I tried a number of names but string is the most descriptive.
